Fix: 1.完善地图页面 2.完善定位->商户列表请求逻辑：如果跳转到商户列表页面没有定位数据会等待定位成功之后请求商户列表数据或者是定位失败直接请求商户列表数据

+++ MaintenanceCar/Models/SCLocationInfo.m
@@ -7,7 +7,6 @@
 //
 
 #import "SCLocationInfo.h"
-#import "transform.h"
 
 static SCLocationInfo *locationInfo = nil;
 
@@ -59,11 +58,11 @@ static SCLocationInfo *locationInfo = nil;
         NSInteger integer   = (NSInteger)distance / 1000;
         if (remainder > 500)
         {
-            displayDistance = [NSString stringWithFormat:@"%ldkm", integer +1 ];
+            displayDistance = [NSString stringWithFormat:@"%ldkm", (long)(integer +1)];
         }
         else
         {
-            displayDistance = [NSString stringWithFormat:@"%ldkm", integer];
+            displayDistance = [NSString stringWithFormat:@"%ldkm", (long)integer];
         }
     }
     return displayDistance;
